Reworked it again.  Removed duplicate functionality and updated alarmStateEvent to better handle all situations.
Added testMode banner to HTML to help people identify when they are in test Mode

Removed some of the testMode logic, made it unpredictable.  Left it in on Refresh so hopefully we won't see false positives like we used to. Downside it might take a few tries to cause the DH to enter test mode.  Another downside is that it will refresh state on next pull from Nest. Think it is a minor issue, may look at it in the future.  I don't want to leave anything in that would prevent the DH from working.